Replace crack with multi_json
The crack gem uses an extracted copy of an outdated Rails JSON parser,
which uses YAML as a backend. This is vulnerable to arbitrary code
execution in certain contexts:

https://groups.google.com/forum/#!topic/rubyonrails-security/1h2DR63ViGo

This patch replaces crack with multi_json, which intelligently selects
from the locally available JSON backends, placing minimal constraints on
client dependencies.