Autoload files instead of require.
I'm aware that most usage of this gem will be to shorten a link using the client and won't need OAuth or many of the other objects. To reduce memory consumption I've chosen to use Kernel#autoload. There have been concerns in the past that autoload isn't threadsafe, but it has been since Ruby 2.0.